+++ src/main/java/humbledude/lotto/domain/AccountManager.java
@@ -1,4 +1,4 @@
-package humbledude.lotto;
+package humbledude.lotto.domain;
 
 public class AccountManager {
