+++ src/test/java/humbledude/lotto/domain/AutoLottoTest.java
@@ -1,4 +1,4 @@
-package humbledude.lotto;
+package humbledude.lotto.domain;
 
 import org.junit.jupiter.api.Test;
