+++ src/test/java/humbledude/lotto/domain/LottoNumbersTest.java
@@ -1,11 +1,6 @@
-package humbledude.lotto;
+package humbledude.lotto.domain;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
-import org.junit.jupiter.params.provider.ValueSource;
-
-import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
